fix: move mood arrows hook to a separate file
